d3d12; Factorise IALayout generation code
